Check if the folder exists in WrappedGitRepo cleanup
Cloning a repository might fail for different reasons, one of them is
when the remote repository does not exist. In such case, it does not
make sense to do any cleanup, because the folder simply does not exist.

When this happens, FileNotFoundError is raised, which does not clearly
represents the real error behind.

Applying this change will skip the cleanup if the folder does not exist,
and the operation will end up failing and providing the real error at
the end when all attempts are done.